refactored env test wrappers
