--- cases/env-type/00-external/implementations/as/index.ts
@@ -1,11 +1,0 @@
-import {
-  Env,
-  Args_externalEnvMethod,
-  ModuleBase
-} from "./wrap";
-
-export class Module extends ModuleBase {
-  externalEnvMethod(args: Args_externalEnvMethod, env: Env): Env {
-    return env;
-  }
-}
