--- cases/env-type/00-external/implementations/rs/lib.rs
@@ -1,10 +1,0 @@
-pub mod wrap;
-pub use wrap::*;
-use wrap::env::{Env};
-use wrap::module::{ModuleTrait, Module};
-
-impl ModuleTrait for Module {
-    fn external_env_method(_: ArgsExternalEnvMethod, env: Env) -> Result<Env, String> {
-        Ok(env)
-    }
-}
