--- cases/env-type/01-main/polywrap.json
@@ -1,10 +1,0 @@
-{
-  "source": {
-    "import_abis": [
-      {
-        "uri": "ens/external-env.polywrap.eth",
-        "abi": "../../../00-external"
-      }
-    ]
-  }
-}
