--- cases/env-type/02-subinvoked/polywrap.json
@@ -1,6 +1,0 @@
-{
-  "source": {
-    "import_abis": [
-    ]
-  }
-}
