--- cases/env-type/02-subinvoked/schema.graphql
@@ -1,7 +1,0 @@
-type Env {
-  str: String!
-}
-
-type Module {
-  envMethod: String! @env(required: true)
-}
