+++ cases/env-type/02-subinvoker-with-env/schema.graphql
@@ -1,0 +1,29 @@
+#import { Module, Env } into Subinvoked from "mock/main"
+
+type EnvObject {
+  prop: String!
+}
+
+enum EnvEnum {
+  FIRST
+  SECOND
+}
+
+type Env {
+  str: String!
+  optStr: String
+  optFilledStr: String
+  number: Int8!
+  optNumber: Int8
+  bool: Boolean!
+  optBool: Boolean
+  en: EnvEnum!
+  optEnum: EnvEnum
+  object: EnvObject!
+  optObject: EnvObject
+  array: [UInt32!]!
+}
+
+type Module {
+  subinvokeMethodRequireEnv: Subinvoked_Env! @env(required: true)
+}
