--- cases/env-type/03-subinvoke/implementations/as/index.ts
@@ -1,13 +1,0 @@
-import {
-  Subinvoked_Module,
-  Args_subinvokeEnvMethod,
-  ModuleBase,
-} from "./wrap";
-
-export class Module extends ModuleBase {
-  subinvokeEnvMethod(args: Args_subinvokeEnvMethod): string {
-    const result = Subinvoked_Module.envMethod({}).unwrap()
-
-    return result;
-  }
-}
