--- cases/env-type/03-subinvoke/polywrap.json
@@ -1,10 +1,0 @@
-{
-  "source": {
-    "import_abis": [
-      {
-        "uri": "ens/subinvoked-env.polywrap.eth",
-        "abi": "../../../02-subinvoked"
-      }
-    ]
-  }
-}
