--- cases/env-type/03-subinvoke/schema.graphql
@@ -1,5 +1,0 @@
-#import { Module } into Subinvoked from "ens/subinvoked-env.polywrap.eth"
-
-type Module {
-  subinvokeEnvMethod: String!
-}
